+++ notionsnapshot/logger.py
@@ -8,7 +8,7 @@ import cssutils
 
 class LoggingWrapper(logging.LoggerAdapter):
     def process(self, msg, kwargs):
-        # proxy for logs that adds indentation based on the stack depth
+        # proxy for logging module that adds indentation based on the stack depth
         ignored_stack_frames = 8
         indentation_level = len(traceback.extract_stack()) - ignored_stack_frames
         tab = " " * 3
@@ -30,7 +30,7 @@ LOG = LoggingWrapper.get_log()
 
 
 def trace(print_args: bool = True):
-    # used as decorator to log function calls and return values
+    # for @trace decorator to log functions arguments and return values
     def decorator(func):
         @functools.wraps(func)
         def wrapper(*args, **kwargs):
